requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.38
botocore==1.19.38
cache==1.0.3
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-embed-video==1.3.3
django-filter==2.4.0
django-haystack==3.0
django-js-asset==1.2.2
django-machina==1.1.3
django-mptt==0.11.0
django-storages==1.11
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
jmespath==0.10.0
markdown2==2.3.10
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
Whoosh==2.7.4
zipp==3.4.0
